RAG_Code/requirements.txt
streamlit
# sentence-transformers
# faiss-cpu==1.7.2
langchain
openai
pymupdf
python-dotenv
langchain-community
chromadb
langchain_groq